bindings: Update bindings for Vertigo 2 & Remove left-handed teleport binding
Note: There seems to be a bug with the left-handed weapon teleport function in Vertigo 2.
To avoid confusion the binding was removed from the "etee_controller_profile.json", but the binding json is still stored in the bindings/ folder, in case we can re-add it when Vertigo 2's team fixes the bug.