fix(battery): Fix battery percentage going to incorrect values for a single frame
This introduces a most common element ring buffer for battery percentages on each controller. This ring buffer keeps track of the battery for a window of 128 frames per controller as an unsigned integer. This allows it to work well for handling the case of battery percentages reporting the incorrect value for a single frame mid-session, triggering a battery low notification.